tests/wpt/web-platform-tests/element-timing/supported-element-type.html

<!DOCTYPE html>
<head>
<title>PerformanceObserver.supportedEntryTypes contains "element"</title>
</head>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  if (typeof PerformanceObserver.supportedEntryTypes === "undefined")
    assert_unreached("supportedEntryTypes is not supported.");
  assert_greater_than(PerformanceObserver.supportedEntryTypes.indexOf("element"), -1,
    "There should be an entry 'element' in PerformanceObserver.supportedEntryTypes");
}, "supportedEntryTypes contains 'element'.");
</script>
</body>
</html>
